% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administration_route_get_dict.R
\name{administration_route_get_dict}
\alias{administration_route_get_dict}
\title{FUNCTION_TITLE}
\usage{
administration_route_get_dict()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Helper function to get the administration route dictionary form the CvT database
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
