% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administration_route_get_unique.R
\name{administration_route_get_unique}
\alias{administration_route_get_unique}
\title{administration_route_get_unique}
\usage{
administration_route_get_unique(fileList, template_path)
}
\arguments{
\item{fileList}{PARAM_DESCRIPTION}

\item{template_path}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:select]{select}, \link[dplyr:mutate]{mutate}, \link[dplyr:bind]{bind_rows}, \link[dplyr:distinct]{distinct}, \link[dplyr:mutate-joins]{left_join}, \link[dplyr:filter]{filter}
\link[writexl:write_xlsx]{write_xlsx}
}
