% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administration_route_update_dict.R
\name{administration_route_update_dict}
\alias{administration_route_update_dict}
\title{administration_route_update_dict}
\usage{
administration_route_update_dict(dict_file)
}
\arguments{
\item{dict_file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[readxl:read_excel]{read_xlsx}
\link[dplyr:select]{select}, \link[dplyr:mutate]{mutate}, \link[dplyr:filter]{filter}, \link[dplyr:filter-joins]{anti_join}
}
