% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convert_failed.R
\name{check_convert_failed}
\alias{check_convert_failed}
\title{check_convert_failed}
\usage{
check_convert_failed(x, f, col, log_path, id_col = "id")
}
\arguments{
\item{x}{Input list of datasets being processed.}

\item{f}{Optional filename for logging purposes.}

\item{col}{String of name of column to check.}

\item{log_path}{File path where to save the log file.}

\item{id_col}{Column to use to log index value to help with logging.}
}
\value{
Modified input \code{x} dataframe list with new "convert_failed" dataframe.
}
\description{
Check if a conversion failed, resulting in an NA value. This can be due to a lack of necessary equation parameters (e.g., molecular weight, subject weight).
}
\seealso{
\link[dplyr:filter]{filter}
}
