% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\title{FUNCTION_TITLE}
\usage{
check_missing(x, miss_col, f, flag = TRUE, log_path)
}
\arguments{
\item{x}{Input list of datasets being normalized}

\item{miss_col}{The column to check}

\item{f}{Filename for flagging purposes}

\item{flag}{Whether to log a flag for the missing column #'}

\item{log_path}{File path where to save the log file.}
}
\value{
Modified version of the input \code{x} parameter
}
\description{
Helper function to check for missing values for desired column.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}
}
