% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_units.R
\name{check_missing_units}
\alias{check_missing_units}
\title{check_missing_units}
\usage{
check_missing_units(x, f, units_col, log_path, flag = TRUE)
}
\arguments{
\item{x}{Input list of datasets being processed.}

\item{f}{Optional filename for logging purposes.}

\item{units_col}{String of the name of the column to check.}

\item{log_path}{File path where to save the log file.}

\item{flag}{Whether to log a flag for the missing column.}
}
\value{
Modified input \code{x} dataframe list with new "missing_units" dataframe.
}
\description{
Function to check for missing units for input column.
}
\seealso{
\link[dplyr:filter]{filter}
}
