% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_non_numeric.R
\name{check_non_numeric}
\alias{check_non_numeric}
\title{check_non_numeric}
\usage{
check_non_numeric(x, f, col)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{col}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:mutate]{mutate}, \link[dplyr:filter]{filter}, \link[dplyr:select]{select}
}
