% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_radiolabel.R
\name{check_radiolabel}
\alias{check_radiolabel}
\title{check_radiolabel}
\usage{
check_radiolabel(raw, f)
}
\arguments{
\item{raw}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:mutate]{mutate}
}
