% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_required_fields.R
\name{check_required_fields}
\alias{check_required_fields}
\title{check_required_fields}
\usage{
check_required_fields(df, f)
}
\arguments{
\item{df}{List of dataframes for the sheets within an extraction template.}

\item{f}{Optional filename for logging purposes.}
}
\value{
None. Logs any flags.
}
\description{
Function to check if processed template is missing required fields.
}
\seealso{
\link[dplyr:filter]{filter}
}
