% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_subject_list.R
\name{check_subject_list}
\alias{check_subject_list}
\title{check_subject_list}
\usage{
check_subject_list(x, f, col, log_path)
}
\arguments{
\item{x}{Input list of datasets being processed.}

\item{f}{Optional filename for logging purposes.}

\item{col}{String of the name of the column to check.}

\item{log_path}{File path where to save the log file.}
}
\value{
Modified input \code{x} dataframe list with new "split_subject" dataframe.
}
\description{
Function to check if an input field has a ; separated list, therefore should be split.
}
\seealso{
\link[dplyr:filter]{filter}
}
