% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unit_ci.R
\name{check_unit_ci}
\alias{check_unit_ci}
\title{check_unit_ci}
\usage{
check_unit_ci(x, f, col, log_path)
}
\arguments{
\item{x}{Input list of datasets being processed.}

\item{f}{Optional filename for logging purposes.}

\item{col}{String of the name of the column to check.}

\item{log_path}{PARAM_DESCRIPTION}
}
\value{
Modified input \code{x} dataframe list with new "ci" dataframe.
}
\description{
Function to check for confidence intervals for an input field.
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:mutate]{mutate}, \link[dplyr:across]{across}
\link[tidyr:reexports]{all_of}
}
