% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unit_range.R
\name{check_unit_range}
\alias{check_unit_range}
\title{check_unit_range}
\usage{
check_unit_range(x, f, col, log_path)
}
\arguments{
\item{x}{Input list of datasets being processed.}

\item{f}{Optional filename for logging purposes.}

\item{col}{String of the name of the column to check.}

\item{log_path}{File path where to save the log file.}
}
\value{
Modified input \code{x} dataframe list with new "unit_range" dataframe.
}
\description{
Function to check for ranges for input column.
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:mutate]{mutate}, \link[dplyr:rowwise]{rowwise}, \link[dplyr:across]{across}, \link[dplyr:select]{select}
\link[tidyr:separate]{separate}, \link[tidyr:reexports]{all_of}
}
