% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_curation_get_curation_files.R
\name{chemical_curation_get_curation_files}
\alias{chemical_curation_get_curation_files}
\title{chemical_curation_get_curation_files}
\usage{
chemical_curation_get_curation_files(f_name)
}
\arguments{
\item{f_name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[readxl:excel_sheets]{excel_sheets}, \link[readxl:read_excel]{read_xlsx}
\link[dplyr:filter]{filter}
}
