% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_curation_match_curated_chemicals.R
\name{chemical_curation_match_curated_chemicals}
\alias{chemical_curation_match_curated_chemicals}
\title{FUNCTION_TITLE}
\usage{
chemical_curation_match_curated_chemicals(df, f_name)
}
\arguments{
\item{df}{Input chemical data to match}

\item{f_name}{File path to curated chemicals (multiple sheets)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:mutate]{mutate}, \link[dplyr:mutate-joins]{left_join}, \link[dplyr:distinct]{distinct}, \link[dplyr:bind_rows]{bind_rows}, \link[dplyr:select]{select}, \link[dplyr:arrange]{arrange}
}
