% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_curation_prep_name_match.R
\name{chemical_curation_prep_name_match}
\alias{chemical_curation_prep_name_match}
\title{chemical_curation_prep_name_match}
\usage{
chemical_curation_prep_name_match(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:mutate]{mutate}
\link[stringr:str_extract]{str_extract_all}
}
