% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical_update_dict.R
\name{chemical_update_dict}
\alias{chemical_update_dict}
\title{chemical_update_dict}
\usage{
chemical_update_dict(dict_file = NULL, dat = NULL)
}
\arguments{
\item{dict_file}{PARAM_DESCRIPTION, Default: NULL}

\item{dat}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[readxl:read_excel]{read_xlsx}
}
