% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemicals_get_dict.R
\name{chemicals_get_dict}
\alias{chemicals_get_dict}
\title{FUNCTION_TITLE}
\usage{
chemicals_get_dict()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Helper function to get the concentration medium dictionary form the CvT database
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
