% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_download_docs.R
\name{clowder_download_docs}
\alias{clowder_download_docs}
\title{Download Clowder Documents by File ID and Filename}
\usage{
clowder_download_docs(docData = NULL, outputDir, apiKey, limit = NULL)
}
\arguments{
\item{docData}{Dataframe containing fileID and fileName columns for the Clowder API download}

\item{outputDir}{PARAM_DESCRIPTION}

\item{apiKey}{The API key required for a user to access the Clowder dataset}

\item{limit}{A numeric integer value to specify the number of documents to pull from the list of file IDs (Leave NULL if you want all to be downloaded)}

\item{fileName}{A character list of file names (with file extension) to name the downloaded files}
}
\value{
None. Files are downloaded into specified output directory
}
\description{
This is a helper function to download documents using the Clowder API. It also checks if a fileID/fileName has already been downloaded and skips if it exists
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[downloader:download]{download}
}
