% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_get_dataset_files.R
\name{clowder_get_dataset_files}
\alias{clowder_get_dataset_files}
\title{clowder_get_dataset_files}
\usage{
clowder_get_dataset_files(dsID, baseurl, apiKey)
}
\arguments{
\item{dsID}{Clowder dataset identifier.}

\item{baseurl}{Clowder base URL.}

\item{apiKey}{Clowder API key.}
}
\value{
Dataframe with Clowder file identifier, folder name, and file name.
}
\description{
Pull a dataframe of file metadata stored in a Clowder dataset.
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{content_type}}, \code{\link[httr]{add_headers}}, \code{\link[httr]{content}}
\code{\link[tidyr]{unnest}}
\code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{select}}
}
