% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_get_dataset_folders.R
\name{clowder_get_dataset_folders}
\alias{clowder_get_dataset_folders}
\title{FUNCTION_TITLE}
\usage{
clowder_get_dataset_folders(dsID, baseurl, apiKey)
}
\arguments{
\item{dsID}{Clowder dataset identifier.}

\item{baseurl}{Clowder base URL.}

\item{apiKey}{Clowder API key.}
}
\value{
Dataframe with Clowder folder identifier and folder name.
}
\description{
Pull a dataframe of folder metadata stored in a Clowder dataset.
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{content_type}}, \code{\link[httr]{add_headers}}, \code{\link[httr]{content}}
\code{\link[tidyr]{unnest}}
\code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{select}}
}
