% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_get_docList_2.R
\name{clowder_get_docList_2}
\alias{clowder_get_docList_2}
\title{Get Clowder Document List (varient of the other function in utilities)}
\usage{
clowder_get_docList_2(dsID = NULL, apiKey = NULL)
}
\arguments{
\item{dsID}{Clowder dataset ID to pull from.}

\item{apiKey}{The API key required for a user to access the Clowder dataset}
}
\value{
Returns a dataframe with file details of: filename and ClowderID.
}
\description{
This is a helper function to get a list of documents available in a Clowder dataset
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[httr:GET]{GET}, \link[httr:content]{content}
\link[tidyr:reexports]{tibble}
\link[purrr:map]{map_chr}
\link[dplyr:select]{select}
}
