% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_get_file_metadata.R
\name{clowder_get_file_metadata}
\alias{clowder_get_file_metadata}
\title{clowder_get_file_metadata}
\usage{
clowder_get_file_metadata(fileID, baseurl, apiKey)
}
\arguments{
\item{fileID}{Vector of Clowder file identifiers.}

\item{baseurl}{Clowder base URL.}

\item{apiKey}{Clowder API key.}
}
\value{
Dataframe with Clowder file metadata.
}
\description{
Pull a dataframe of folder metadata stored in a Clowder dataset.
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{content_type}}, \code{\link[httr]{add_headers}}, \code{\link[httr]{content}}
\code{\link[purrr]{pluck}}, \code{\link[purrr]{keep}}
\code{\link[tidyr]{unnest}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{bind_cols}}, \code{\link[dplyr]{bind_rows}}
}
