% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_match_docs.R
\name{clowder_match_docs}
\alias{clowder_match_docs}
\title{clowder_match_docs}
\usage{
clowder_match_docs(
  df = NULL,
  dsID = NULL,
  baseurl = NULL,
  apiKey = NULL,
  clowder_file_list = NULL
)
}
\arguments{
\item{df}{Input template document's sheet for mapping.}

\item{dsID}{Clowder dataset ID to pull from.}

\item{baseurl}{Clowder base URL.}

\item{apiKey}{API key to access Clowder repo.}

\item{clowder_file_list}{Opitonal input list of Clowder files from clowder_get_dataset_files().}
}
\value{
Modified \code{df} with clowder_file_id field for matched documents.
}
\description{
Function to match template documents to Clowder files.
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:mutate]{mutate}, \link[dplyr:mutate-joins]{left_join}, \link[dplyr:arrange]{arrange}
}
