% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_match_post_upload.R
\name{clowder_match_post_upload}
\alias{clowder_match_post_upload}
\title{FUNCTION_TITLE}
\usage{
clowder_match_post_upload(dsID = NULL, apiKey = NULL)
}
\arguments{
\item{dsID}{Clowder dataset ID to pull from.}

\item{apiKey}{The API key required for a user to access the Clowder dataset}
}
\value{
Returns a dataframe with file details of: filename and ClowderID.
}
\description{
This is a helper function to match already loaded CvT document entries to Clowder docs.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:select]{select}
}
