% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_upload_file_metadata.R
\name{clowder_upload_file_metadata}
\alias{clowder_upload_file_metadata}
\title{clowder_upload_file_metadata}
\usage{
clowder_upload_file_metadata(metadata, dsID, userID, baseurl, apiKey)
}
\arguments{
\item{metadata}{Dataframe of metadata to add to Clowder files.}

\item{dsID}{Clowder dataset identifier.}

\item{userID}{Clowder User identifier.}

\item{baseurl}{Clowder base URL.}

\item{apiKey}{Clowder API key.}
}
\value{
None. Clowder APi calls are performed to add metadata to Clowder files.
}
\description{
Function to process add metadata field values to Clowder files.
}
\seealso{
\code{\link[tidyr]{unite}}, \code{\link[tidyr]{separate}}
\code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{filter}}
\code{\link[purrr]{flatten}}
\code{\link[httr]{POST}}, \code{\link[httr]{content_type}}, \code{\link[httr]{add_headers}}
}
