% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cols_to_NA.R
\name{convert_cols_to_NA}
\alias{convert_cols_to_NA}
\title{FUNCTION_TITLE}
\usage{
convert_cols_to_NA(df, col_list)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{col_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
A helper function to convert input dataframe column list to NA
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
