% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mass_per_mass.R
\name{convert_mass_per_mass}
\alias{convert_mass_per_mass}
\title{FUNCTION_TITLE}
\usage{
convert_mass_per_mass(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Helpfer function to help convert mass/mass concentrations
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[readxl:read_excel]{read_xlsx}
\link[dplyr:select]{select}, \link[dplyr:mutate]{mutate}, \link[dplyr:rename]{rename}, \link[dplyr:mutate-joins]{left_join}, \link[dplyr:across]{across}
}
