% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{convert_units}
\usage{
convert_units(
  x,
  num,
  units,
  desired,
  conv_factor = NA,
  overwrite_units = FALSE
)
}
\arguments{
\item{x}{Input dataframe to convert.}

\item{num}{Name of column with values to convert.}

\item{units}{Name of column with units to convert from.}

\item{desired}{Desired units to convert the input value into.}

\item{conv_factor}{Conversion factor to use (such as Molecular weight, tissue Density, etc.).}

\item{overwrite_units}{Boolean to overwrite the 'units' with desired units.}
}
\value{
Modified dataframe of input \code{x} dataframe with converted column using \code{convert_get_conversion_factor()}.
}
\description{
Function to convert input values to desired units.
}
