% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units_grepl.R
\name{convert_units_grepl}
\alias{convert_units_grepl}
\title{convert_units_grepl}
\usage{
convert_units_grepl(unit_type)
}
\arguments{
\item{unit_type}{Input unit type (e.g., weight, height, age, dose_duration, conc)}
}
\value{
List of unit name standardizations by input unit type.
}
\description{
Function to get various grepl statements for unit name standardization.
}
