% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect_to_CvT.R
\name{db_connect_to_CvT}
\alias{db_connect_to_CvT}
\title{db_connect_to_CvT}
\usage{
db_connect_to_CvT()
}
\value{
A database connection object.
}
\description{
A function to create a connection to the CvT database using the .Renviron file parameters.
}
\seealso{
\link[DBI:dbConnect]{dbConnect}, \link[RPostgres:Postgres]{PostgreSQL}
}
