% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_tbl_id.R
\name{db_get_tbl_id}
\alias{db_get_tbl_id}
\title{db_get_tbl_id}
\usage{
db_get_tbl_id(tblName = NULL, idFilter = NULL)
}
\arguments{
\item{tblName}{The name of the table to pull the ID from.}

\item{idFilter}{A SQL WHERE statement to filter idName column to. If empty, pulls all data.}
}
\value{
A list of ID values from the specified database table.
}
\description{
A function to pull table fk identification from a specified table by a SQL filter statement, or the entire table.
}
\seealso{
\link[DBI:dbSendQuery]{dbSendQuery}, \link[DBI:dbClearResult]{dbClearResult}, \link[DBI:dbDisconnect]{dbDisconnect}
\link[DBI:dbFetch]{dbFetch}
}
