% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_initialize_CvTdb.R
\name{db_initialize_CvTdb}
\alias{db_initialize_CvTdb}
\title{FUNCTION_TITLE}
\usage{
db_initialize_CvTdb()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function to write empty tables for CvTdb SQLite file from template docs.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
