% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_initialize_CvTdb_from_RDat.R
\name{db_initialize_CvTdb_from_RDat}
\alias{db_initialize_CvTdb_from_RDat}
\title{FUNCTION_TITLE}
\usage{
db_initialize_CvTdb_from_RDat(file = NULL)
}
\arguments{
\item{file}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function to write empty tables for CvTdb SQLite file from template docs.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
