% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_push_tbl_to_db.R
\name{db_push_tbl_to_db}
\alias{db_push_tbl_to_db}
\title{db_push_tbl_to_db}
\usage{
db_push_tbl_to_db(
  dat = NULL,
  tblName = NULL,
  fieldTypes = NULL,
  overwrite = FALSE,
  customSQL = NULL,
  append = FALSE
)
}
\arguments{
\item{dat}{A dataframe to write to the database, Default: NULL}

\item{tblName}{Name of database table to create or append and fill with input \code{dat} dataframe data, Default: NULL}

\item{fieldTypes}{Named list of field types for columns, Default: NULL}

\item{overwrite}{Boolean of whether to overwrite the \code{tblName} table with input \code{dat} dataframe data, Default: FALSE}

\item{customSQL}{Optional custom SQL statement to push, Default: NULL}

\item{append}{Boolean of whether to append the \code{tblName} table with input \code{dat} dataframe data, Default: FALSE}
}
\value{
None. Updates are pushed to the database.
}
\description{
A function to push a dataframe to a specified table in the database.
}
\seealso{
\link[DBI:dbWriteTable]{dbWriteTable}, \link[DBI:dbSendQuery]{dbSendQuery}, \link[DBI:dbDisconnect]{dbDisconnect}
}
