% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_update_tbl.R
\name{db_update_tbl}
\alias{db_update_tbl}
\title{db_update_tbl}
\usage{
db_update_tbl(df = NULL, tblName = NULL)
}
\arguments{
\item{df}{A dataframe to write to the database, Default: NULL}

\item{tblName}{Name of database table to create or append and fill with input \code{df} dataframe data, Default: NULL}
}
\value{
None. Updates are pushed to the database.
}
\description{
A function to update database table entries based on dataframe
}
\seealso{
\link[DBI:dbWriteTable]{dbWriteTable}, \link[DBI:dbClearResult]{dbClearResult}, \link[DBI:dbDisconnect]{dbDisconnect}
\link[DBI:dbSendStatement]{dbSendStatement}
}
