% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_query_cvt.R
\name{db_query_cvt}
\alias{db_query_cvt}
\title{db_query_cvt}
\usage{
db_query_cvt(query = NULL, query_type = "query")
}
\arguments{
\item{query}{A SQL query string.}

\item{query_type}{Input string specifying if the query is a SQL "query" (retrieve data) or "statement" (update database). Default: 'query'.}
}
\value{
Return objects from DBI \code{dbGetQuery()} and \code{dbSendStatement()} functions.
}
\description{
Function to query or send statements to the database. Handles errors/warnings with tryCatch.
}
\seealso{
\link[DBI:dbGetQuery]{dbGetQuery}, \link[DBI:dbSendStatement]{dbSendStatement}, \link[DBI:dbDisconnect]{dbDisconnect}
}
