% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_frequency_get_dict.R
\name{dose_frequency_get_dict}
\alias{dose_frequency_get_dict}
\title{FUNCTION_TITLE}
\usage{
dose_frequency_get_dict()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
#'@description Helper function to get the concentration medium dictionary form the CvT database

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
