% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_jira_update_clowder_info.R
\name{download_jira_update_clowder_info}
\alias{download_jira_update_clowder_info}
\title{download_jira_update_clowder_info}
\usage{
download_jira_update_clowder_info(
  jira_project,
  in_file = NULL,
  auth_token,
  reset_attachments = FALSE,
  update_clowder_metadata = FALSE,
  dsID,
  baseurl,
  userID,
  apiKey,
  labels_filter = NULL,
  epic_filter = c(),
  attachment_filter = c()
)
}
\arguments{
\item{jira_project}{Jira project identifier (e.g., CVTDB).}

\item{in_file}{Optional param for the file path to previously pulled Jira information CSV file.}

\item{auth_token}{Jira API token.}

\item{reset_attachments}{Boolean whether to re-download Jira ticket attachments. Default: FALSE.}

\item{update_clowder_metadata}{Boolean whether to update Clowder file metadata. Only set to TRUE once Jira ticket attachments have been uploaded to Clowder. Default: FALSE.}

\item{dsID}{Clowder dataset identifier.}

\item{baseurl}{Clowder base URL.}

\item{userID}{Clowder user identifier.}

\item{apiKey}{Clowder API key.}

\item{labels_filter}{Vector list of Jira ticket labels to filter to.}

\item{epic_filter}{Custom filtering to a specific ticket Epic link by name (single or vector). Default: empty vector.}

\item{attachment_filter}{Filename regex string vector to filter to select Jira ticket attachments, Default: empty vector.}
}
\value{
Dataframe of metadata from Jira tickets to associate to Clowder files.
}
\description{
Function to generate a dataframe of Jira ticket metadata with attachment information. Can be used to download the attachments locally and upload metadata to Clowder files once they have been uploaded to Clowder separately.
}
\seealso{
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}
\code{\link[tidyr]{unite}}
\code{\link[utils]{View}}, \code{\link[utils]{download.file}}
}
