% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_units.R
\name{extract_units}
\alias{extract_units}
\title{extract_units}
\usage{
extract_units(x, units_col, conv_col, unit_type)
}
\arguments{
\item{x}{Input dataframe.}

\item{units_col}{String of the name of the column containing units information with missing values.}

\item{conv_col}{String of the name of the column to try to extract unit information from.}

\item{unit_type}{Input unit type (e.g., weight, height, age, dose_duration, conc), used by \code{convert_units_grepl()}.}
}
\value{
Modified version of input \code{x} dataframe where the input \code{units_col} will have filled in units values from input \code{conv_col} or be "missing_units" tagged instead of NA.
}
\description{
Generic function to extract units from input columns. The goal is to try to fill in missing unit values from \code{units_col} based on string values in \code{conv_col}.
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:mutate]{mutate}, \link[dplyr:across]{across}, \link[dplyr:bind_rows]{bind_rows}
\link[dplyr:reexports]{all_of}
}
