% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cvt_push_ready.R
\name{get_cvt_push_ready}
\alias{get_cvt_push_ready}
\title{FUNCTION_TITLE}
\usage{
get_cvt_push_ready()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function to return a dataframe of files ready to push due to all
flags being '0'.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[readxl:read_excel]{read_xlsx}
\link[dplyr:filter]{filter}, \link[dplyr:across]{across}
}
