% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cvtdb_sqlite.R
\name{get_cvtdb_sqlite}
\alias{get_cvtdb_sqlite}
\title{Get CvTdb SQLite}
\usage{
get_cvtdb_sqlite(schema, outdir)
}
\arguments{
\item{schema}{Database postgreSQL schema}
}
\value{
None. SQLite file is written to output/release folder
}
\description{
Pull CvTdb tables and write sqlite file
}
