% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mw_chemicals_api.R
\name{get_mw_chemicals_api}
\alias{get_mw_chemicals_api}
\title{get_mw_chemicals_api}
\usage{
get_mw_chemicals_api(dtxsid_list, api_key)
}
\arguments{
\item{dtxsid_list}{List of DTXSIDs to query.}

\item{api_key}{API key for the CCTE Chemicals API.}
}
\value{
Dataframe of molecular weight information by DTXSID.
}
\description{
Function to get molecular weight dictionary from CCTE Chemicals API in batches.
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{content}}, \code{\link[httr]{POST}}, \code{\link[httr]{content_type}}, \code{\link[httr]{add_headers}}
\code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{filter}}
}
