% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_tbl_id.R
\name{get_next_tbl_id}
\alias{get_next_tbl_id}
\title{get_next_tbl_id}
\usage{
get_next_tbl_id(schema)
}
\arguments{
\item{schema}{PostgreSQL database schema}
}
\description{
Function to get the next ID increment value for schema tables
}
