% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jira_download_templates.R
\name{jira_download_templates}
\alias{jira_download_templates}
\title{jira_download_templates}
\usage{
jira_download_templates(in_data, auth_token)
}
\arguments{
\item{in_data}{Input dataframe with Jira ticket attachment information.}

\item{auth_token}{Jira API token.}
}
\value{
None. Jira API calls are made to download files to the "output" folder.
}
\description{
Function to download Jira ticket attachments.
}
\seealso{
\code{\link[tidyr]{unite}}
\code{\link[dplyr]{mutate}}
\code{\link[utils]{download.file}}, \code{\link[utils]{unzip}}
}
