% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_database_sheet_group.R
\name{load_database_sheet_group}
\alias{load_database_sheet_group}
\title{FUNCTION_TITLE}
\usage{
load_database_sheet_group(docID, template_path)
}
\arguments{
\item{docID}{The CvT document ID for the file of interest}

\item{template_path}{The file path for the extraction template. If not supplied, hard coded columns will be used.}
}
\value{
A dataframe of the combined sheets
}
\description{
A function to load and pull all sheets from files already loaded to CvTdb.
It corrects for missing required column names from a template file by filling with NA values.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[readxl:read_excel]{read_xlsx}, \link[readxl:excel_sheets]{excel_sheets}, \link[readxl:read_excel]{read_excel}
\link[dplyr:rename]{rename}, \link[dplyr:mutate-joins]{left_join}, \link[dplyr:select]{select}, \link[dplyr:mutate]{mutate}, \link[dplyr:select_all]{rename_all}, \link[dplyr:filter]{filter}, \link[dplyr:distinct]{distinct}
\link[tidyr:reexports]{all_of}
}
