% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file_from_api.R
\name{load_file_from_api}
\alias{load_file_from_api}
\title{load_file_from_api}
\usage{
load_file_from_api(url, headers, file_type, mode = "w")
}
\arguments{
\item{url}{URL where the file will be downloaded.}

\item{headers}{Optional headers to complete the download call.}

\item{file_type}{String of the file extension. Currently only supports "csv" and "xlsx".}

\item{mode}{Download file mode, Default: 'w' for "write".}
}
\value{
Dataframe or list of dataframes from a file downloaded from a URL.
}
\description{
Function to load a CSV or XLSX file as a dataframe or list of dataframes using an API call. Typically, this is to load a CvTdb template file from Clowder.
}
\seealso{
\code{\link[utils]{download.file}}
\code{\link[readr]{read_delim}}, \code{\link[readr]{cols}}
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}
\code{\link[stringr]{str_trim}}
\code{\link[readxl]{excel_sheets}}, \code{\link[readxl]{read_excel}}
}
