% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_CvT_doc_load.R
\name{log_CvT_doc_load}
\alias{log_CvT_doc_load}
\title{log_CvT_doc_load}
\usage{
log_CvT_doc_load(
  f,
  m = NULL,
  reset = FALSE,
  val = NULL,
  log_path = "output/template_normalization_log.xlsx"
)
}
\arguments{
\item{f}{Filename to flag.}

\item{m}{Log field name, Default: NULL.}

\item{reset}{Boolean to reset a row's flags, Default: FALSE.}

\item{val}{Custom field value. Default: NULL.}

\item{log_path}{File path where to save the log file. Default "output/template_normalization_log.xlsx".}
}
\value{
None. Log file is updated.
}
\description{
Function to write a log entry for a template or dataframe being processed.
}
\seealso{
\link[readxl:read_excel]{read_xlsx}
\link[stats:setNames]{setNames}
\link[writexl:write_xlsx]{write_xlsx}
}
