% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_age_category.R
\name{map_age_category}
\alias{map_age_category}
\title{map_age_category}
\usage{
map_age_category(x, dict)
}
\arguments{
\item{x}{Input dataframe with species and age value/units fields.}

\item{dict}{Input dataframe with species age category assignment fields.}
}
\value{
Modified input \code{x} dataframe with age_category field filled in where possible.
}
\description{
Function to use a standard age category dictionary file to map species age to an age category.
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:select]{select}
}
