% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_chemical_fk.R
\name{match_chemical_fk}
\alias{match_chemical_fk}
\title{FUNCTION_TITLE}
\usage{
match_chemical_fk(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Helper function TBD.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
