% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_cvt_doc_to_db_doc.R
\name{match_cvt_doc_to_db_doc}
\alias{match_cvt_doc_to_db_doc}
\title{match_cvt_doc_to_db_doc}
\usage{
match_cvt_doc_to_db_doc(df = NULL)
}
\arguments{
\item{df}{Input dataframe of Document records to try to match to database Documents table entries, Default: NULL.}
}
\value{
Modified input \code{df} dataframe with matched database Documents table ID values.
}
\description{
Function to match input document metadata to CvTdb document records by pmid, other_study_identifier, doi, url hierarchy.
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{tidyeval-compat}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{arrange}}
\code{\link[stringr]{str_trim}}
}
