% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_subject_fk.R
\name{match_subject_fk}
\alias{match_subject_fk}
\title{FUNCTION_TITLE}
\usage{
match_subject_fk(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
A helper function to match a new study entry to existing studies entries (if available).
In development, waiting to figure out bset way to match given string entries.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:select]{select}
}
