% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_extrapolate.R
\name{norm_extrapolate}
\alias{norm_extrapolate}
\title{FUNCTION_TITLE}
\usage{
norm_extrapolate(x, f, extrap_type)
}
\arguments{
\item{x}{Input list of datasets being normalized}

\item{f}{Filename for flagging purposes}

\item{extrap_type}{The type of extrapolation being performed (weight is the only accepted form at this time). #'}
}
\value{
Modified version of the input \code{x} parameter
}
\description{
Helper function to check for missing metric and attemp to exrapolate values.
Eventually make this a generic function for all norm extrapolation processes...
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:mutate]{mutate}, \link[dplyr:across]{across}, \link[dplyr:group_by]{group_by}, \link[dplyr:summarise]{summarise}, \link[dplyr:select]{select}, \link[dplyr:mutate-joins]{left_join}, \link[dplyr:distinct]{distinct}, \link[dplyr:rename]{rename}
}
