% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_age.R
\name{normalize_age}
\alias{normalize_age}
\title{normalize_age}
\usage{
normalize_age(raw, f, log_path, debug = FALSE)
}
\arguments{
\item{raw}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{log_path}{File path where to save the log file.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[readxl:read_excel]{read_xlsx}
\link[dplyr:mutate]{mutate}, \link[dplyr:filter]{filter}, \link[dplyr:select]{select}, \link[dplyr:bind_rows]{bind_rows}, \link[dplyr:rename]{rename}, \link[dplyr:arrange]{arrange}
}
