% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_age_units.R
\name{normalize_age_units}
\alias{normalize_age_units}
\title{normalize_age_units}
\usage{
normalize_age_units(x)
}
\arguments{
\item{x}{Input vector of conc units.}
}
\value{
Modified input \code{x} vector with normalized age unit names.
}
\description{
Helper function to normalize age unit names.
}
