% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_boolean.R
\name{normalize_boolean}
\alias{normalize_boolean}
\title{normalize_boolean}
\usage{
normalize_boolean(x, col)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{col}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
