% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_conc.R
\name{normalize_conc}
\alias{normalize_conc}
\title{FUNCTION_TITLE}
\usage{
normalize_conc(raw, f)
}
\arguments{
\item{raw}{A dataframe of weight information to normalize}

\item{f}{The file name of the template being processed. Used for error logging. #'}
}
\value{
Normalized version of the input \code{raw} parameter.
}
\description{
A helper function to normalize concentration units.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:mutate]{mutate}, \link[dplyr:filter]{filter}, \link[dplyr:across]{across}, \link[dplyr:bind]{bind_rows}, \link[dplyr:arrange]{arrange}
\link[httk:get_physchem_param]{get_physchem_param}
}
