% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_conc_medium.R
\name{normalize_conc_medium}
\alias{normalize_conc_medium}
\title{normalize_conc_medium}
\usage{
normalize_conc_medium(raw, f)
}
\arguments{
\item{raw}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[readxl:read_excel]{read_xlsx}
\link[dplyr:mutate]{mutate}, \link[dplyr:rename]{rename}, \link[dplyr:select]{select}, \link[dplyr:mutate-joins]{left_join}, \link[dplyr:filter]{filter}
}
