% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_conc_units.R
\name{normalize_conc_units}
\alias{normalize_conc_units}
\title{FUNCTION_TITLE}
\usage{
normalize_conc_units(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Helpfer function to normalize concentration units
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
