% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_CvT_data.R
\name{normalize_CvT_data}
\alias{normalize_CvT_data}
\title{FUNCTION_TITLE}
\usage{
normalize_CvT_data(df, f, log_path)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{log_path}{File path where to save the log file.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
A helper function to orchestrate normalization of CvT Data
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:mutate-joins]{left_join}, \link[dplyr:select]{select}, \link[dplyr:mutate]{mutate}, \link[dplyr:distinct]{distinct}
}
