% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_dose.R
\name{normalize_dose}
\alias{normalize_dose}
\title{FUNCTION_TITLE}
\usage{
normalize_dose(raw, f)
}
\arguments{
\item{raw}{A dataframe of weight information to normalize}

\item{f}{The file name of the template being processed. Used for error logging. #'}
}
\value{
Normalized version of the input \code{raw} parameter.
}
\description{
Doses are stored in mg/kg. In cases where the mass of the subject was not reported, and the dose was
administered as a simple mass and not a body weight proportion, a mg/kg dose was calculated using the average
mass for all subjects of that type.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:mutate]{mutate}, \link[dplyr:bind]{bind_rows}, \link[dplyr:arrange]{arrange}, \link[dplyr:select]{select}
\link[httk:get_physchem_param]{get_physchem_param}
}
