% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_species.R
\name{normalize_species}
\alias{normalize_species}
\title{normalize_species}
\usage{
normalize_species(x, log_path = NULL)
}
\arguments{
\item{x}{Dataframe with species information to normalize.}

\item{log_path}{File path where to save the log file.}
}
\value{
Modified input \code{x} dataframe with normalized species information.
}
\description{
Normalize species field for input dataframe.
}
