% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_time.R
\name{normalize_time}
\alias{normalize_time}
\title{normalize_time}
\usage{
normalize_time(raw, f)
}
\arguments{
\item{raw}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:mutate]{mutate}, \link[dplyr:bind]{bind_rows}, \link[dplyr:arrange]{arrange}
}
