% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_time_units.R
\name{normalize_time_units}
\alias{normalize_time_units}
\title{normalize_time_units}
\usage{
normalize_time_units(x)
}
\arguments{
\item{x}{Input vector of time units.}
}
\value{
Modified \code{x} vector with normalized time unit names.
}
\description{
Function to normalize time unit names to a standard format.
}
