% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_weight.R
\name{normalize_weight}
\alias{normalize_weight}
\title{FUNCTION_TITLE}
\usage{
normalize_weight(raw, f)
}
\arguments{
\item{raw}{A dataframe of weight information to normalize}

\item{f}{The file name of the template being processed. Used for error logging. #'}
}
\value{
Normalized version of the input \code{raw} parameter.
}
\description{
A helper function to normalize subject weight (associated with dose normalization).
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:mutate]{mutate}, \link[dplyr:filter]{filter}, \link[dplyr:bind]{bind_rows}, \link[dplyr:arrange]{arrange}, \link[dplyr:select]{select}
}
