% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_clowder_files_to_load.R
\name{pull_clowder_files_to_load}
\alias{pull_clowder_files_to_load}
\title{pull_clowder_files_to_load
Function to pull Clowder templates to process based on metadata "cvt_to_load"}
\usage{
pull_clowder_files_to_load(
  dsID,
  baseurl,
  apiKey,
  curation_set_tag,
  metadata_filter_tag = NULL
)
}
\description{
pull_clowder_files_to_load
Function to pull Clowder templates to process based on metadata "cvt_to_load"
}
