% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_jira_info.R
\name{pull_jira_info}
\alias{pull_jira_info}
\title{pull_jira_info}
\usage{
pull_jira_info(
  jira_project,
  in_file = NULL,
  auth_token = NULL,
  status_filter = "Done",
  epic_filter = c()
)
}
\arguments{
\item{jira_project}{Jira project code (e.g. CVTDB).}

\item{auth_token}{Authorization token for Jira.}

\item{status_filter}{Custom filtering to a ticket status or vector of statuses. Default: NULL.}

\item{epic_filter}{Custom filtering to a specific ticket Epic link by name (single or vector). Default: empty vector.}

\item{download_bulk}{Boolean whether to bulk download ticket attachments, Default: FALSE.}
}
\value{
Summary DataFrame of Jira tickets by Epic, Label, and Status.
}
\description{
Script to process CSV export of Jira into a status log
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 out = pull_jira_info(jira_project="project_name")
 }
}
}
\seealso{
\link[utils:download.file]{download.file}, \link[utils:unzip]{unzip}
\link[readr:read_delim]{read_csv}, \link[readr:cols]{cols}
\link[dplyr:select]{select}, \link[dplyr:reexports]{contains}, \link[dplyr:mutate]{mutate}, \link[dplyr:reexports]{everything}, \link[dplyr:filter]{filter}, \link[dplyr:distinct]{distinct}, \link[dplyr:mutate-joins]{left_join}, \link[dplyr:group_by]{group_by}, \link[dplyr:summarise]{summarise}, \link[dplyr:context]{n}
\link[tidyr:unite]{unite}
\link[stringr:str_trim]{str_squish}
}
