% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_add_record.R
\name{qc_add_record}
\alias{qc_add_record}
\title{qc_add_record}
\usage{
qc_add_record(df, tbl_field_list, load_doc_sheet_only)
}
\arguments{
\item{df}{Input dataframe of field values to add to database table}

\item{tbl_name}{Name of table the records are from}
}
\value{
Modified version of input df with mapped new database ID values
}
\description{
Function to add record from QC template
}
