% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_remove_record.R
\name{qc_remove_record}
\alias{qc_remove_record}
\title{qc_remove_record}
\usage{
qc_remove_record(df, tbl_name)
}
\arguments{
\item{df}{Input dataframe of id, qc_notes, and qc_flags for records to remove}

\item{tbl_name}{Name of table the records are from}
}
\value{
None.
}
\description{
Function to remove QC record flagged for removal. Function will account
for needed cascade of removal for foreign key table connections
}
