% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_normalized_template.R
\name{save_normalized_template}
\alias{save_normalized_template}
\title{FUNCTION_TITLE}
\usage{
save_normalized_template(df = doc_sheet_list, f = f)
}
\arguments{
\item{df}{PARAM_DESCRIPTION, Default: doc_sheet_list}

\item{f}{PARAM_DESCRIPTION, Default: f}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
A helper function to cache the normalized templates
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[tools:fileutils]{file_ext}
\link[writexl:write_xlsx]{write_xlsx}
}
