% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_original_fields.R
\name{set_original_fields}
\alias{set_original_fields}
\title{set_original_fields}
\usage{
set_original_fields(sheet_list, schema)
}
\arguments{
\item{sheet_list}{Dataframe list to rename with original columns}

\item{schema}{String for the PostgreSQL schema information to pull.}
}
\description{
Pull dictionary of original fields from database tables and rename in sheet_list
}
