% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_get_unique_to_curate.R
\name{species_get_unique_to_curate}
\alias{species_get_unique_to_curate}
\title{species_get_unique_to_curate}
\usage{
species_get_unique_to_curate(fileList, template_path)
}
\arguments{
\item{fileList}{Optional list of template files to load. If NULL, queries the database to check for cases.}

\item{template_path}{Path to input CvTdb template file.}
}
\value{
Console output of species cases to add to \code{normalize_species()}.
}
\description{
Load a list of templates or query the database to check for species cases to add to \code{normalize_species()}.
}
\seealso{
\link[dplyr:select]{select}
}
