% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.dictionary.entries.from.file.R
\name{update.dictionary.entries.from.file}
\alias{update.dictionary.entries.from.file}
\title{update.dictionary.entries.from.file}
\usage{
\method{update}{dictionary.entries.from.file}(schema, in_file)
}
\arguments{
\item{schema}{Database schema for PostgreSQL.}

\item{in_file}{Input file with dictionary updates to process.}
}
\value{
Dataframe list of dictionaries.
}
\description{
Function to update dictionary entries from file generated from
get.dictionary.entries.to.curate().
}
\seealso{
\code{\link[readxl]{excel_sheets}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}
}
