% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_CvT_data.R
\name{validate_cvt}
\alias{validate_cvt}
\title{validate_cvt}
\usage{
validate_cvt(
  clowder_file_id = NULL,
  clowder_api_key = NULL,
  file_path = NULL,
  db_identifier = NULL,
  df = NULL,
  df_identifier = NULL,
  log_path = "output/validation/validate_cvt_log.xlsx",
  ignore_present = FALSE,
  ignore_nonempty = FALSE,
  ignore_required = FALSE,
  ignore_qc = FALSE,
  ignore_field_types = FALSE,
  ignore_field_entries = FALSE,
  ignore_field_uniqueness = FALSE,
  ignore_foreign_keys = FALSE,
  ignore_expected_fields = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{clowder_file_id}{Clowder template file identifier, Default: NULL}

\item{clowder_api_key}{Clowder API token, Default: NULL}

\item{file_path}{Path to template file, Default: NULL}

\item{db_identifier}{String or numeric identifier for a record in the database Documents table, Default: NULL}

\item{df}{Input named list of dataframes, Default: NULL}

\item{df_identifier}{String identifier for input \code{df}, required if \code{df} is not NULL, Default: NULL}

\item{log_path}{Path to log, Default: 'output/validation/validate_cvt_log.xlsx'}

\item{ignore_present}{Boolean of whether to ignore \code{validate_sheets_present()}, Default: FALSE}

\item{ignore_nonempty}{Boolean of whether to ignore \code{validate_sheets_nonempty()}, Default: FALSE}

\item{ignore_required}{Boolean of whether to ignore \code{validate_required_fields()}, Default: FALSE}

\item{ignore_qc}{Boolean of whether to ignore \code{validate_qc_fields()} and/or use the QC template for validation, Default: FALSE}

\item{ignore_field_types}{Boolean of whether to ignore \code{validate_field_types()}, Default: FALSE}

\item{ignore_field_entries}{Boolean of whether to ignore \code{validate_field_entries()}, Default: FALSE}

\item{ignore_field_uniqueness}{Boolean of whether to ignore \code{validate_field_uniqueness()}, Default: FALSE}

\item{ignore_foreign_keys}{Boolean of whether to ignore \code{validate_foreign_keys()}, Default: FALSE}

\item{ignore_expected_fields}{Boolean of whether to ignore \code{validate_expected_fields()}, Default: FALSE}

\item{verbose}{Boolean of whether to print additional console messages, Default: FALSE.}
}
\value{
Boolean of whether the input template passed the validation.
}
\description{
Function to validate a CvTdb template based on a ruleset.
}
\seealso{
\code{\link[dplyr]{mutate}}
}
