% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_expected_fields.R
\name{validate_expected_fields}
\alias{validate_expected_fields}
\title{Check expected fields}
\usage{
validate_expected_fields(df, f, log_path, template_path, verbose = FALSE)
}
\arguments{
\item{df}{List of dataframes for the sheets within an extraction template}

\item{f}{Filename for flagging purposes}

\item{log_path}{File path where to save the log file. Default "output/template_normalization_log.xlsx"}

\item{template_path}{File path to the location of a blank template for comparison}
}
\value{
None. Logs any flags
}
\description{
Function to check if processed document's column names appropriately match the blank template.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[validate:validator]{validator}
}
