% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_field_types.R
\name{validate_field_types}
\alias{validate_field_types}
\title{Validate Field Types}
\usage{
validate_field_types(df, f, log_path, verbose = FALSE)
}
\arguments{
\item{df}{List of dataframes for the sheets within an extraction template}

\item{f}{Filename for flagging purposes}

\item{log_path}{File path where to save the log file. Default "output/template_normalization_log.xlsx"}
}
\value{
Boolean of document validity. Logs any errors
}
\description{
Function to check if processed document has invalid field types.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[validate:validator]{validator}
}
