% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_foreign_keys.R
\name{validate_foreign_keys}
\alias{validate_foreign_keys}
\title{Check foreign keys validator}
\usage{
validate_foreign_keys(df, f, log_path, verbose = FALSE)
}
\arguments{
\item{df}{List of dataframes for the sheets within an extraction template}

\item{f}{Filename for flagging purposes}

\item{log_path}{File path where to save the log file. Default "output/template_normalization_log.xlsx"}
}
\value{
None. Logs any flags
}
\description{
Function to check if processed document's foreign keys match to a value in their respective sheets.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[validate:validator]{validator}
}
