% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_qc_fields.R
\name{validate_qc_fields}
\alias{validate_qc_fields}
\title{validate_qc_fields}
\usage{
validate_qc_fields(df, f, log_path, verbose = FALSE)
}
\arguments{
\item{df}{Input named list of dataframes.}

\item{f}{Filename for logging purposes.}

\item{log_path}{Path to log.}

\item{verbose}{Boolean of whether to print additional console messages, Default: FALSE.}
}
\value{
Boolean of whether the input template passed the validation.
}
\description{
Function to check if processed QC template was filled out appropriately.
}
\seealso{
\code{\link[validate]{validator}}, \code{\link[validate]{confront}}, \code{\link[validate]{summary}}, \code{\link[validate]{meta}}, \code{\link[validate]{satisfying}}
\code{\link[dplyr]{filter}}
}
