% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sheets_nonempty.R
\name{validate_sheets_nonempty}
\alias{validate_sheets_nonempty}
\title{validate_sheets_nonempty}
\usage{
validate_sheets_nonempty(df, f, log_path)
}
\arguments{
\item{df}{Input named list of dataframes.}

\item{f}{Filename for logging purposes.}

\item{log_path}{Path to log.}
}
\value{
Boolean of whether the input template passed the validation.
}
\description{
Function to check if processed template contains expected non-empty sheets.
}
