% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sheets_present.R
\name{validate_sheets_present}
\alias{validate_sheets_present}
\title{validate_sheets_present}
\usage{
validate_sheets_present(df, f, log_path)
}
\arguments{
\item{df}{Input named list of dataframes.}

\item{f}{Filename for logging purposes.}

\item{log_path}{Path to log.}
}
\value{
Boolean of whether the input template passed the validation.
}
\description{
Validate if template contains expected sheets
}
