% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convert_failed.R
\name{check_convert_failed}
\alias{check_convert_failed}
\title{check_convert_failed}
\usage{
check_convert_failed(x, f, col, log_path, id_col = "id")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{col}{PARAM_DESCRIPTION}

\item{log_path}{File path where to save the log file.}

\item{id_col}{Column to use to log index value}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}
}
