% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\title{check_missing}
\usage{
check_missing(x, miss_col, f, flag = TRUE, log_path)
}
\arguments{
\item{x}{Input list of datasets being processed.}

\item{miss_col}{String of the name of the column to check.}

\item{f}{Optional filename for logging purposes.}

\item{flag}{Whether to log a flag for the missing column.}

\item{log_path}{File path where to save the log file.}
}
\value{
Modified input \code{x} dataframe list with new "missing" dataframe.
}
\description{
Function to check for missing values for input column.
}
\seealso{
\link[dplyr:filter]{filter}
}
