% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_units.R
\name{check_missing_units}
\alias{check_missing_units}
\title{FUNCTION_TITLE}
\usage{
check_missing_units(x, f, units_col, log_path, flag = TRUE)
}
\arguments{
\item{x}{Input list of datasets being normalized}

\item{f}{Filename for flagging purposes}

\item{units_col}{The units column to check #'}

\item{log_path}{File path where to save the log file.}

\item{flag}{Whether to log a flag for the missing column #'}
}
\value{
Modified version of the input \code{x} parameter
}
\description{
Helper function to check for missing units for desired metric column.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}
}
