% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_non_numeric.R
\name{check_non_numeric}
\alias{check_non_numeric}
\title{check_non_numeric}
\usage{
check_non_numeric(x, f, col, log_path)
}
\arguments{
\item{x}{Input list of datasets being processed.}

\item{f}{Optional filename for logging purposes.}

\item{col}{String of the name of the column to check.}

\item{log_path}{File path where to save the log file.}
}
\value{
Modified input \code{x} dataframe list with new "non_numeric" dataframe.
}
\description{
Function to check for non-numeric values for input column.
}
\seealso{
\link[dplyr:mutate]{mutate}, \link[dplyr:filter]{filter}, \link[dplyr:select]{select}
}
