% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_required_fields.R
\name{check_required_fields}
\alias{check_required_fields}
\title{FUNCTION_TITLE}
\usage{
check_required_fields(df, f)
}
\arguments{
\item{df}{List of dataframes for the sheets within an extraction template}

\item{f}{Filename for flagging purposes #'}
}
\value{
None. Logs any flags
}
\description{
Function to check if processed document is missing required fields.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}
}
