% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_required_fields.R
\name{validate_required_fields}
\alias{validate_required_fields}
\title{Check required fields validator}
\usage{
validate_required_fields(df, f, log_path, verbose = FALSE)
}
\arguments{
\item{df}{List of dataframes for the sheets within an extraction template}

\item{f}{Filename for flagging purposes}

\item{log_path}{File path where to save the log file. Default "output/template_normalization_log.xlsx"}
}
\value{
None. Logs any flags
}
\description{
Function to check if processed document is missing required fields.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[validate:validator]{validator}
}
