% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_subject_list.R
\name{check_subject_list}
\alias{check_subject_list}
\title{FUNCTION_TITLE}
\usage{
check_subject_list(x, f, col, log_path)
}
\arguments{
\item{x}{Input list of datasets being normalized}

\item{f}{Filename for flagging purposes}

\item{col}{The column being checked/normalized #'}

\item{log_path}{File path where to save the log file.}
}
\value{
Modified version of the input \code{x} parameter
}
\description{
Helper function to check if a metric entry has a ; separated list, therefore should be split.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}
}
