% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unit_range.R
\name{check_unit_range}
\alias{check_unit_range}
\title{FUNCTION_TITLE}
\usage{
check_unit_range(x, f, col, estimated)
}
\arguments{
\item{x}{Input list of datasets being normalized}

\item{f}{Filename for flagging purposes}

\item{col}{The column being checked/normalized}

\item{estimated}{The column for *_estimated flags (0, 1, 2) #'}
}
\value{
Modified version of the input \code{x} parameter
}
\description{
Helper function to check for ranges, and handling them
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:mutate]{mutate}, \link[dplyr:rowwise]{rowwise}, \link[dplyr:across]{across}, \link[dplyr:select]{select}
\link[tidyr:separate]{separate}, \link[tidyr:reexports]{all_of}
}
