% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_match_post_upload.R
\name{clowder_match_post_upload}
\alias{clowder_match_post_upload}
\title{clowder_match_post_upload}
\usage{
clowder_match_post_upload(dsID = NULL, baseurl = NULL, apiKey = NULL)
}
\arguments{
\item{dsID}{Clowder dataset identifier.}

\item{baseurl}{Clowder base URL.}

\item{apiKey}{Clowder API key.}
}
\value{
None. An update query is performed to update the Documents table "clowder_file_id" field.
}
\description{
This is a helper function to match already loaded CvT document entries to Clowder docs.
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:select]{select}
}
