% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cols_to_NA.R
\name{convert_cols_to_NA}
\alias{convert_cols_to_NA}
\title{convert_cols_to_NA}
\usage{
convert_cols_to_NA(df, col_list)
}
\arguments{
\item{df}{Input dataframe.}

\item{col_list}{String of name of column to convert to NA.}
}
\value{
Modified input \code{df} dataframe with \code{col_list} as NA.
}
\description{
A function to convert input dataframe column list to NA.
}
