% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{FUNCTION_TITLE}
\usage{
convert_units(x, num, units, desired, MW = NA, overwrite_units = FALSE)
}
\arguments{
\item{x}{Input dataframe to convert}

\item{num}{Name of column with values to convert}

\item{units}{Name of column with units to convert from}

\item{desired}{Desired units to convert the input value into}

\item{MW}{Molecular weight of the input chemical (if applicable)}

\item{overwrite_units}{Boolean to overwrite the 'units' with desired units.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
A helper function to convert input values to desired units.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
