% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units_grepl.R
\name{convert_units_grepl}
\alias{convert_units_grepl}
\title{convert_units_grepl}
\usage{
convert_units_grepl(unit_type)
}
\arguments{
\item{unit_type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function to get various grepl statements for unit extraction/conversion
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
