% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect_to_CvT.R
\name{db_connect_to_CvT}
\alias{db_connect_to_CvT}
\title{FUNCTION_TITLE}
\usage{
db_connect_to_CvT()
}
\value{
A database connection object
}
\description{
A function to create a connection to the CvT database
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[RPostgreSQL:dbConnect]{dbConnect}, \link[RPostgreSQL:PostgreSQL]{PostgreSQL}
}
