% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_tbl_id.R
\name{db_get_tbl_id}
\alias{db_get_tbl_id}
\title{FUNCTION_TITLE}
\usage{
db_get_tbl_id(tblName = NULL, idFilter = NULL)
}
\arguments{
\item{tblName}{PARAM_DESCRIPTION, Default: NULL}

\item{idFilter}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
A list of ID values from the specified database table

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
#'@description A function to pull table fk identification from a specified table
by a SQL filter statement, or the entire table.

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[RPostgres:dbSendQuery]{dbSendQuery}, \link[RPostgres:dbClearResult]{dbClearResult}, \link[RPostgres:dbDisconnect]{dbDisconnect}
\link[DBI:dbFetch]{dbFetch}
}
