% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_push_tbl_to_db.R
\name{db_push_tbl_to_db}
\alias{db_push_tbl_to_db}
\title{db_push_tbl_to_db}
\usage{
db_push_tbl_to_db(
  dat = NULL,
  tblName = NULL,
  fieldTypes = NULL,
  overwrite = FALSE,
  customSQL = NULL,
  append = FALSE
)
}
\arguments{
\item{dat}{PARAM_DESCRIPTION, Default: NULL}

\item{tblName}{PARAM_DESCRIPTION, Default: NULL}

\item{fieldTypes}{PARAM_DESCRIPTION, Default: NULL}

\item{overwrite}{PARAM_DESCRIPTION, Default: FALSE}

\item{customSQL}{PARAM_DESCRIPTION, Default: NULL}

\item{append}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[RPostgreSQL:dbWriteTable]{dbWriteTable}, \link[RPostgreSQL:dbSendQuery]{dbSendQuery}, \link[RPostgreSQL:dbDisconnect]{dbDisconnect}
}
