% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_push_to_CvT.R, R/db_update_tbl.R
\name{db_push_to_CvT}
\alias{db_push_to_CvT}
\alias{db_update_tbl}
\title{FUNCTION_TITLE}
\usage{
db_push_to_CvT(df = NULL, tblName = NULL)

db_update_tbl(df = NULL, tblName = NULL)
}
\arguments{
\item{df}{A dataframe to write to the database}

\item{tblName}{The name of the table to write the df data to}
}
\value{
None

None
}
\description{
A function to push a dataframe to a specified table in a database

A function to update database table entries based on dataframe
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[RPostgres:dbWriteTable]{dbWriteTable}, \link[RPostgres:dbClearResult]{dbClearResult}, \link[RPostgres:dbDisconnect]{dbDisconnect}
\link[DBI:dbSendStatement]{dbSendStatement}

\link[DBI:dbWriteTable]{dbWriteTable}, \link[DBI:dbClearResult]{dbClearResult}, \link[DBI:dbDisconnect]{dbDisconnect}
\link[DBI:dbSendStatement]{dbSendStatement}
}
