% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_query_cvt.R
\name{db_query_cvt}
\alias{db_query_cvt}
\title{FUNCTION_TITLE}
\usage{
db_query_cvt(query = NULL)
}
\arguments{
\item{query}{A SQL query string to query the database with}
}
\value{
Dataframe of database query results
}
\description{
A helper function to query cvt and receive the results.
Handles errors/warnings with tryCatch.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[RPostgres:dbGetQuery]{dbGetQuery}, \link[RPostgres:dbDisconnect]{dbDisconnect}
}
