% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_jira_update_clowder_info.R
\name{download_jira_update_clowder_info}
\alias{download_jira_update_clowder_info}
\title{download_jira_update_clowder_info}
\usage{
download_jira_update_clowder_info(
  jira_project,
  in_file = NULL,
  auth_token,
  reset_attachments = FALSE,
  update_clowder_metadata = FALSE,
  dsID,
  baseurl,
  userID,
  apiKey,
  labels_filter = NULL,
  epic_filter = c(),
  attachment_filter = c()
)
}
\arguments{
\item{jira_project}{Jira project identifier}

\item{in_file}{Optional param for the file path to previously pulled Jira information}

\item{auth_token}{Jira API token}

\item{reset_attachments}{Boolean whether to re-download files, or just update Clowder metadata}

\item{dsID}{Clowder dataset ID}

\item{baseurl}{Clowder URL}

\item{userID}{Clowder user ID}

\item{apiKey}{Clowder API key}

\item{labels_filter}{Vector list of Jira ticket labels to filter to}

\item{epic_filter}{Custom filtering to a specific ticket Epic link by name (single or vector). Default: empty vector.}
}
\description{
download_jira_update_clowder_info
}
