% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_units.R
\name{extract_units}
\alias{extract_units}
\title{FUNCTION_TITLE}
\usage{
extract_units(x, units_col, conv_col, unit_type)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{units_col}{PARAM_DESCRIPTION}

\item{conv_col}{PARAM_DESCRIPTION}

\item{unit_type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Generic function to extract units from input columns
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:mutate]{mutate}, \link[dplyr:across]{across}, \link[dplyr:bind_rows]{bind_rows}
\link[tidyr:reexports]{all_of}
}
