% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dictionary.entries.to.curate.R
\name{get.dictionary.entries.to.curate}
\alias{get.dictionary.entries.to.curate}
\title{get.dictionary.entries.to.curate}
\usage{
get.dictionary.entries.to.curate(schema, full.report = FALSE)
}
\arguments{
\item{schema}{Database schema for PostgreSQL}
}
\value{
Dataframe list of dictionaries
}
\description{
Function to generate file of dictionary entries to curate
normalized field values. Includes foreign key identifiers to help query database
}
