% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cvt_template.R
\name{get_cvt_template}
\alias{get_cvt_template}
\title{get_cvt_template}
\usage{
get_cvt_template(template_path)
}
\arguments{
\item{template_path}{Path to the CvT template file.}
}
\value{
Named list of template file dataframes.
}
\description{
Load the CvT template file into a list of empty dataframes.
}
\seealso{
\link[readxl:excel_sheets]{excel_sheets}, \link[readxl:read_excel]{read_xlsx}
}
