% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dict_update_ids.R
\name{get_dict_update_ids}
\alias{get_dict_update_ids}
\title{get_dict_update_ids}
\usage{
get_dict_update_ids(sheet_list, schema)
}
\arguments{
\item{sheet_list}{Dataframe list to rename with original columns}

\item{schema}{String for the PostgreSQL schema information to pull.}
}
\description{
Update database dictionaries and match foreign keys
}
