% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mw_chemicals_api.R
\name{get_mw_chemicals_api}
\alias{get_mw_chemicals_api}
\title{get_mw_chemicals_api
Function to get MW dictionary from CCTE Chemicals API in batches}
\usage{
get_mw_chemicals_api(dtxsid_list, api_key)
}
\arguments{
\item{dtxsid_list}{List of DTXSIDs to query}

\item{api_key}{API key for the CCTE Chemicals API}
}
\value{
dataframe of MW information by DTXSID
}
\description{
get_mw_chemicals_api
Function to get MW dictionary from CCTE Chemicals API in batches
}
