% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_CvT_doc_load.R
\name{log_CvT_doc_load}
\alias{log_CvT_doc_load}
\title{log_CvT_doc_load}
\usage{
log_CvT_doc_load(f, m = NULL, reset = FALSE)
}
\arguments{
\item{f}{PARAM_DESCRIPTION}

\item{m}{PARAM_DESCRIPTION, Default: NULL}

\item{reset}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[readxl:read_excel]{read_xlsx}
\link[stats:setNames]{setNames}
\link[writexl:write_xlsx]{write_xlsx}
}
