% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_age_category.R
\name{map_age_category}
\alias{map_age_category}
\title{map_age_category}
\usage{
map_age_category(x, dict)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{dict}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:select]{select}
}
