% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_age_units.R
\name{normalize_age_units}
\alias{normalize_age_units}
\title{normalize_age_units}
\usage{
normalize_age_units(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
