% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_boolean.R
\name{normalize_boolean}
\alias{normalize_boolean}
\title{normalize_boolean}
\usage{
normalize_boolean(x, col)
}
\arguments{
\item{x}{Input dataframe}

\item{col}{Stirng vector of fields to update boolean values to "0" or "1".}
}
\value{
Modified dataframe \code{x} with boolean fields updated to "0" or "1".
}
\description{
FUNCTION_DESCRIPTION
}
