% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_conc.R
\name{normalize_conc}
\alias{normalize_conc}
\title{normalize_conc}
\usage{
normalize_conc(raw, f, log_path, debug = FALSE)
}
\arguments{
\item{raw}{Input dataframe of data with data to normalize.}

\item{f}{Optional filename for logging purposes.}

\item{log_path}{File path where to save the log file.}

\item{debug}{Boolean of whether to stop conversion logic early for debugging purpose. Default: FALSE.}
}
\value{
Normalized version of the input \code{raw} parameter.
}
\description{
A helper function to normalize dose.
}
\seealso{
\link[dplyr:mutate]{mutate}, \link[dplyr:filter]{filter}, \link[dplyr:across]{across}, \link[dplyr:bind_rows]{bind_rows}, \link[dplyr:arrange]{arrange}
\link[httr:POST]{POST}
}
