% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_conc_units.R
\name{normalize_conc_units}
\alias{normalize_conc_units}
\title{normalize_conc_units}
\usage{
normalize_conc_units(x)
}
\arguments{
\item{x}{Input vector of conc units.}
}
\value{
Modified input \code{x} vector with normalized concentration unit names.
}
\description{
Helper function to normalize concentration unit names.
}
