% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_height.R
\name{normalize_height}
\alias{normalize_height}
\title{FUNCTION_TITLE}
\usage{
normalize_height(raw, f, log_path, debug = FALSE)
}
\arguments{
\item{raw}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{log_path}{File path where to save the log file.}
}
\value{
Normalized version of the input \code{raw} parameter.

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
#'@description A helper function to normalize subject height.

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:mutate]{mutate}, \link[dplyr:filter]{filter}, \link[dplyr:bind_rows]{bind_rows}, \link[dplyr:arrange]{arrange}
}
