% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_species.R
\name{normalize_species}
\alias{normalize_species}
\title{normalize_species}
\usage{
normalize_species(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
