% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_time.R
\name{normalize_time}
\alias{normalize_time}
\title{normalize_time}
\usage{
normalize_time(raw, f, log_path, debug = FALSE)
}
\arguments{
\item{raw}{Input dataframe of data with data to normalize.}

\item{f}{Optional filename for logging purposes.}

\item{log_path}{File path where to save the log file.}

\item{debug}{Boolean of whether to stop conversion logic early for debugging purpose. Default: FALSE.}
}
\value{
Modified dataframe of \code{raw} with normalized values.
}
\description{
Normalize time to hours for input dataframe.
}
\seealso{
\link[dplyr:mutate]{mutate}, \link[dplyr:bind_rows]{bind_rows}, \link[dplyr:arrange]{arrange}
}
