% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_time.R
\name{normalize_time}
\alias{normalize_time}
\title{normalize_time}
\usage{
normalize_time(raw, f, log_path, debug = FALSE)
}
\arguments{
\item{raw}{PARAM_DESCRIPTION}

\item{f}{PARAM_DESCRIPTION}

\item{log_path}{File path where to save the log file.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:mutate]{mutate}, \link[dplyr:bind_rows]{bind_rows}, \link[dplyr:arrange]{arrange}
}
