% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_add_record.R
\name{qc_add_record}
\alias{qc_add_record}
\title{qc_add_record}
\usage{
qc_add_record(df, tbl_field_list, load_doc_sheet_only, col_exclude)
}
\arguments{
\item{df}{Input dataframe of field values to add to database table.}

\item{tbl_field_list}{Dataframe of CvT tables and fields.}

\item{load_doc_sheet_only}{Boolean whether just to add document sheet only.}

\item{col_exclude}{List of columns to exclude from database pushes.}
}
\value{
None. SQL statements are submitted to add a record to the database tables.
}
\description{
Function to add record from QC template.
}
\seealso{
\code{\link[purrr]{keep}}
\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}
}
