% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_remove_record.R
\name{qc_remove_record}
\alias{qc_remove_record}
\title{qc_remove_record}
\usage{
qc_remove_record(df, tbl_name, reset_extraction = FALSE, del_qc_note = NULL)
}
\arguments{
\item{df}{Input dataframe of id, qc_notes, and qc_flags for records to remove}

\item{tbl_name}{Name of table the records are from}

\item{reset_extraction}{Boolean whether to reset_extraction or remove whole record. Defaul FALSE.}

\item{del_qc_note}{Optional note to add to qc_notes and qc_flags field for tbl_name records.}
}
\value{
None. SQL statements are run to delete records in a cascading fashion if they have foreign key linkages to core tables.
}
\description{
Function to remove QC record flagged for removal. Function will account
for needed cascade of removal for foreign key table connections
}
\seealso{
\code{\link[purrr]{keep}}
\code{\link[dplyr]{mutate}}
}
