% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorganize_file_flags.R
\name{reorganize_file_flags}
\alias{reorganize_file_flags}
\title{reorganize_file_flags}
\usage{
reorganize_file_flags()
}
\value{
None. File are reorganized in output directory.
}
\description{
A function to check if a file has logged issues (changed to
1 for select columns) and move it to appropriate subfolder.
}
\seealso{
\link[readxl:read_excel]{read_xlsx}
}
