% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_normalized_template.R
\name{save_normalized_template}
\alias{save_normalized_template}
\title{save_normalized_template}
\usage{
save_normalized_template(df, f)
}
\arguments{
\item{df}{Named list of dataframes of normalized template data.}

\item{f}{Filename of the input template that is modified to save a normalized copy.}
}
\value{
None. Input dataframe is saved to output folder.
}
\description{
A helper function to cache the normalized templates during the normalization workflow.
}
\seealso{
\link[tools:fileutils]{file_ext}
\link[writexl:write_xlsx]{write_xlsx}
}
