% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_get_unique_to_curate.R
\name{species_get_unique_to_curate}
\alias{species_get_unique_to_curate}
\title{species_get_unique_to_curate}
\usage{
species_get_unique_to_curate(fileList, template_path)
}
\arguments{
\item{fileList}{PARAM_DESCRIPTION}

\item{template_path}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:select]{select}
}
