% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_expected_fields.R
\name{validate_expected_fields}
\alias{validate_expected_fields}
\title{validate_expected_fields}
\usage{
validate_expected_fields(df, f, log_path, template_path, verbose = FALSE)
}
\arguments{
\item{df}{Input named list of dataframes.}

\item{f}{Filename for logging purposes.}

\item{log_path}{Path to log.}

\item{template_path}{Path to blank template file for comparison.}

\item{verbose}{Boolean of whether to print additional console messages, Default: FALSE.}
}
\value{
Boolean of whether the input template passed the validation.
}
\description{
Function to check if processed document's column names appropriately match the blank template.
}
\seealso{
\link[validate:validator]{validator}
}
