% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_field_uniqueness.R
\name{validate_field_uniqueness}
\alias{validate_field_uniqueness}
\title{validate_field_uniqueness}
\usage{
validate_field_uniqueness(df, f, log_path, verbose = FALSE)
}
\arguments{
\item{df}{Input named list of dataframes.}

\item{f}{Filename for logging purposes.}

\item{log_path}{Path to log.}

\item{verbose}{Boolean of whether to print additional console messages, Default: FALSE.}
}
\value{
Boolean of whether the input template passed the validation.
}
\description{
Function to check if processed document has unique fields where applicable.
}
\seealso{
\link[validate:validator]{validator}
}
