% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_foreign_keys.R
\name{validate_foreign_keys}
\alias{validate_foreign_keys}
\title{validate_foreign_keys}
\usage{
validate_foreign_keys(df, f, log_path, verbose = FALSE)
}
\arguments{
\item{df}{Input named list of dataframes.}

\item{f}{Filename for logging purposes.}

\item{log_path}{Path to log.}

\item{verbose}{Boolean of whether to print additional console messages, Default: FALSE.}
}
\value{
Boolean of whether the input template passed the validation.
}
\description{
Function to check if processed document's foreign keys match to a value in their respective sheets.
}
\seealso{
\link[validate:validator]{validator}
}
